+++ backend/canisters/community/api/src/updates/claim_prize.rs
@@ -23,6 +23,8 @@ pub enum Response {
     AlreadyClaimed,
     PrizeFullyClaimed,
     PrizeEnded,
+    LedgerError,
     TransferFailed(String, FailedCryptoTransaction),
     FailedAfterTransfer(String, CompletedCryptoTransaction),
+    InternalError(String),
 }