+++ backend/canisters/user/api/src/updates/tip_message.rs
@@ -1,7 +1,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{CanisterId, Chat, CompletedCryptoTransaction, Cryptocurrency, MessageId, MessageIndex, Milliseconds, UserId};
+use types::{CanisterId, Chat, Cryptocurrency, MessageId, MessageIndex, Milliseconds, UserId};
 
 #[ts_export(user, tip_message)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -35,5 +35,5 @@ pub enum Response {
     TooManyFailedPinAttempts(Milliseconds),
     UserSuspended,
     Retrying(String),
-    InternalError(String, Box<CompletedCryptoTransaction>),
+    InternalError(String),
 }