+++ backend/canisters/user/impl/src/crypto.rs
@@ -1,9 +1,10 @@
 use crate::read_state;
+use ic_cdk::api::call::CallResult;
 use types::{CompletedCryptoTransaction, FailedCryptoTransaction, PendingCryptoTransaction, UserId};
 
 pub async fn process_transaction(
     transaction: PendingCryptoTransaction,
-) -> Result<CompletedCryptoTransaction, FailedCryptoTransaction> {
+) -> CallResult<Result<CompletedCryptoTransaction, FailedCryptoTransaction>> {
     process_transaction_internal(transaction, true).await
 }
 
@@ -14,14 +15,14 @@ pub async fn process_transaction(
 // If calling this method, ensure that the caller has been validated earlier on.
 pub async fn process_transaction_without_caller_check(
     transaction: PendingCryptoTransaction,
-) -> Result<CompletedCryptoTransaction, FailedCryptoTransaction> {
+) -> CallResult<Result<CompletedCryptoTransaction, FailedCryptoTransaction>> {
     process_transaction_internal(transaction, false).await
 }
 
 async fn process_transaction_internal(
     transaction: PendingCryptoTransaction,
     check_caller: bool,
-) -> Result<CompletedCryptoTransaction, FailedCryptoTransaction> {
+) -> CallResult<Result<CompletedCryptoTransaction, FailedCryptoTransaction>> {
     let my_user_id = read_state(|state| {
         if check_caller && state.env.caller() != state.data.owner {
             panic!("Only the owner can transfer cryptocurrency");