+++ backend/libraries/ledger_utils/src/lib.rs
@@ -1,4 +1,5 @@
 use candid::Principal;
+use ic_cdk::api::call::CallResult;
 use ic_ledger_types::{AccountIdentifier, Subaccount, DEFAULT_SUBACCOUNT};
 use sha2::{Digest, Sha256};
 use types::{
@@ -34,16 +35,18 @@ pub async fn process_transaction(
     transaction: PendingCryptoTransaction,
     sender: CanisterId,
     retry_if_bad_fee: bool,
-) -> Result<CompletedCryptoTransaction, FailedCryptoTransaction> {
+) -> CallResult<Result<CompletedCryptoTransaction, FailedCryptoTransaction>> {
     match transaction {
         PendingCryptoTransaction::NNS(t) => nns::process_transaction(t, sender).await,
         PendingCryptoTransaction::ICRC1(t) => match icrc1::process_transaction(t, sender, retry_if_bad_fee).await {
-            Ok(c) => Ok(c.into()),
-            Err(f) => Err(f.into()),
+            Ok(Ok(c)) => Ok(Ok(c.into())),
+            Ok(Err(c)) => Ok(Err(c.into())),
+            Err(e) => Err(e),
         },
         PendingCryptoTransaction::ICRC2(t) => match icrc2::process_transaction(t, sender).await {
-            Ok(c) => Ok(c.into()),
-            Err(f) => Err(f.into()),
+            Ok(Ok(c)) => Ok(Ok(c.into())),
+            Ok(Err(c)) => Ok(Err(c.into())),
+            Err(e) => Err(e),
         },
     }
 }